+++ e2e/react-router/scroll-restoration-sandbox-vite/src/routes/index.tsx
@@ -17,6 +17,8 @@ function HomeComponent() {
       </p>
       {(
         [
+          linkOptions({ to: '/issue-7040-source' }),
+          linkOptions({ to: '/issue-7040-target' }),
           linkOptions({ to: '/normal-page' }),
           linkOptions({ to: '/lazy-page' }),
           linkOptions({ to: '/virtual-page' }),