+++ packages/react-start/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tanstack/react-start
 
+## 1.167.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-router@1.168.7
+  - @tanstack/react-start-client@1.166.22
+  - @tanstack/react-start-server@1.166.22
+  - @tanstack/start-client-core@1.167.6
+  - @tanstack/start-plugin-core@1.167.13
+  - @tanstack/start-server-core@1.167.6
+
+## 1.167.11
+
+### Patch Changes
+
+- Updated dependencies [[`5ca661c`](https://github.com/TanStack/router/commit/5ca661c2f8a7e50167b4112c64aa06cef4148ea9)]:
+  - @tanstack/react-router@1.168.6
+  - @tanstack/react-start-client@1.166.21
+  - @tanstack/react-start-server@1.166.21
+
+## 1.167.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-plugin-core@1.167.12
+
+## 1.167.9
+
+### Patch Changes
+
+- Updated dependencies [[`cf5f554`](https://github.com/TanStack/router/commit/cf5f5542476137a81515099ad740747e84512f9a)]:
+  - @tanstack/react-router@1.168.5
+  - @tanstack/react-start-client@1.166.20
+  - @tanstack/react-start-server@1.166.20
+  - @tanstack/start-client-core@1.167.5
+  - @tanstack/start-plugin-core@1.167.11
+  - @tanstack/start-server-core@1.167.5
+
 ## 1.167.8
 
 ### Patch Changes