+++ packages/router-generator/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @tanstack/router-generator
 
+## 1.166.21
+
+### Patch Changes
+
+- Updated dependencies [[`42c3f3b`](https://github.com/TanStack/router/commit/42c3f3b3a3a478fd6d6894310ef94b2d23794b8e)]:
+  - @tanstack/router-core@1.168.6
+
+## 1.166.20
+
+### Patch Changes
+
+- Fix file-based route generation when custom `routeToken` or `indexToken` values start with regex metacharacters like `+`. ([#7048](https://github.com/TanStack/router/pull/7048))
+
+## 1.166.19
+
+### Patch Changes
+
+- Updated dependencies [[`cf5f554`](https://github.com/TanStack/router/commit/cf5f5542476137a81515099ad740747e84512f9a)]:
+  - @tanstack/router-core@1.168.5
+
 ## 1.166.18
 
 ### Patch Changes