+++ packages/solid-start-client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/solid-start-client
 
+## 1.166.20
+
+### Patch Changes
+
+- Updated dependencies [[`42c3f3b`](https://github.com/TanStack/router/commit/42c3f3b3a3a478fd6d6894310ef94b2d23794b8e)]:
+  - @tanstack/router-core@1.168.6
+  - @tanstack/solid-router@1.168.6
+  - @tanstack/start-client-core@1.167.6
+
+## 1.166.19
+
+### Patch Changes
+
+- Updated dependencies [[`cf5f554`](https://github.com/TanStack/router/commit/cf5f5542476137a81515099ad740747e84512f9a)]:
+  - @tanstack/router-core@1.168.5
+  - @tanstack/solid-router@1.168.5
+  - @tanstack/start-client-core@1.167.5
+
 ## 1.166.18
 
 ### Patch Changes