+++ packages/start-static-server-functions/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/start-static-server-functions
 
+## 1.166.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.167.6
+  - @tanstack/react-start@1.167.12
+  - @tanstack/solid-start@1.167.11
+
+## 1.166.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.167.5
+  - @tanstack/react-start@1.167.9
+  - @tanstack/solid-start@1.167.9
+
 ## 1.166.20
 
 ### Patch Changes