+++ packages/vue-start/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @tanstack/vue-start
 
+## 1.167.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.167.6
+  - @tanstack/start-plugin-core@1.167.13
+  - @tanstack/start-server-core@1.167.6
+  - @tanstack/vue-router@1.168.6
+  - @tanstack/vue-start-client@1.166.20
+  - @tanstack/vue-start-server@1.166.20
+
+## 1.167.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-plugin-core@1.167.12
+
+## 1.167.9
+
+### Patch Changes
+
+- Updated dependencies [[`cf5f554`](https://github.com/TanStack/router/commit/cf5f5542476137a81515099ad740747e84512f9a)]:
+  - @tanstack/vue-router@1.168.5
+  - @tanstack/start-client-core@1.167.5
+  - @tanstack/start-plugin-core@1.167.11
+  - @tanstack/start-server-core@1.167.5
+  - @tanstack/vue-start-client@1.166.19
+  - @tanstack/vue-start-server@1.166.19
+
 ## 1.167.8
 
 ### Patch Changes